reset update
