+++ lyrebird/mock/dm/__init__.py
@@ -12,7 +12,6 @@ from jinja2 import Template
 from lyrebird.application import config
 from lyrebird.mock.handlers import snapshot_helper
 from lyrebird.mock.dm.jsonpath import jsonpath
-from lyrebird.mock.dm.file_data_adapter import FileDataAdapter, PropWriter
 
 
 PROP_FILE_NAME = '.lyrebird_prop'
@@ -41,9 +40,6 @@ class DataManager:
         self.save_to_group_id = None
         self.tmp_group = {'id': 'tmp_group', 'type': 'group', 'name': 'tmp-group', 'label': [], 'children': []}
         self.snapshot_helper = snapshot_helper.SnapshotHelper()
-        self.data_save_controller = FileDataAdapter
-        self._save_prop = FileDataAdapter.save_prop
-        self._save_data = FileDataAdapter.save_data
 
     def set_root(self, root_path):
         """
@@ -64,7 +60,7 @@ class DataManager:
             raise RootNotSet
         _root_prop_path = self.root_path / PROP_FILE_NAME
         if not _root_prop_path.exists():
-            self._save_prop(self.root, self.root_path)
+            self._save_prop()
         with codecs.open(_root_prop_path) as f:
             _root_prop = json.load(f)
             self.root = _root_prop
@@ -308,7 +304,7 @@ class DataManager:
         data['rule'] = _data_rule
 
         data_path = (kwargs.get('output') or self.root_path) / _data_id
-        with codecs.open(data_path, 'w') as f: # 
+        with codecs.open(data_path, 'w') as f:
             json.dump(data, f, ensure_ascii=False)
             logger.debug(f'*** Write file {data_path}')
 
@@ -492,7 +488,7 @@ class DataManager:
                 raise DataNotFound(f'File path: {origin_file_path}')
         new_file_id = target_data_node['id']
         new_file_path = self.root_path / new_file_id
-        with codecs.open(origin_file_path, 'r') as inputfile, codecs.open(new_file_path, 'w') as outputfile: # 
+        with codecs.open(origin_file_path, 'r') as inputfile, codecs.open(new_file_path, 'w') as outputfile:
             origin_text = inputfile.read()
             prop = json.loads(origin_text)
             prop['id'] = new_file_id
@@ -503,6 +499,28 @@ class DataManager:
         COPY_NODE_NAME_SUFFIX = ' - copy'
         return _node['name'] + COPY_NODE_NAME_SUFFIX
 
+    def _save_prop(self):
+        self._sort_children_by_name()
+        prop_str = PropWriter().parse(self.root)
+        # Save prop
+        prop_file = self.root_path / PROP_FILE_NAME
+        with codecs.open(prop_file, 'w') as f:
+            f.write(prop_str)
+        # Reactive mock data
+        _activated_group = self.activated_group
+        self.deactivate()
+        for _group_id in _activated_group:
+            self.activate(_group_id)
+
+    def _sort_children_by_name(self):
+        for node_id in self.id_map:
+            node = self.id_map[node_id]
+            if 'children' not in node:
+                # fix mock data group has no children
+                if node['type'] == 'group':
+                    node['children'] = []
+                continue
+            node['children'] = sorted(node['children'], key=lambda sub_node: sub_node['name'])
 
     # -----
     # Conflict checker
@@ -639,7 +657,7 @@ class DataManager:
         data_file = self.root_path / _id
         if not data_file.exists():
             raise DataNotFound(_id)
-        with codecs.open(data_file, 'w') as f: #
+        with codecs.open(data_file, 'w') as f:
             data_str = json.dumps(data, ensure_ascii=False)
             f.write(data_str)
         self._save_prop()
@@ -652,7 +670,7 @@ class DataManager:
         prop_str = PropWriter().parse(node)
 
         prop_file = outfile_path / PROP_FILE_NAME
-        with codecs.open(prop_file, 'w') as f: #
+        with codecs.open(prop_file, 'w') as f:
             f.write(prop_str)
 
     def _write_file_to_custom_path(self, outfile_path, file_content):
@@ -753,6 +771,10 @@ class NoneClipbord(Exception):
     pass
 
 
+class DumpPropError(Exception):
+    pass
+
+
 class NonePropFile(Exception):
     pass
 
@@ -768,3 +790,72 @@ class NodeExist(Exception):
 class SnapshotEventNotInCorrectFormat(Exception):
     pass
 
+
+class PropWriter:
+
+    def __init__(self):
+        self.indent = 0
+        self.parsers = {
+            'dict': self.dict_parser,
+            'list': self.list_parser,
+            'int': self.int_parser,
+            'str': self.str_parser,
+            'bool': self.bool_parser,
+            'NoneType': self.none_parser
+        }
+
+    def parse(self, prop):
+        prop_type = type(prop)
+        parser = self.parsers.get(prop_type.__name__)
+        if not parser:
+            raise DumpPropError(f'Not support type {prop_type}')
+        return parser(prop)
+
+    def dict_parser(self, val):
+        dict_str = '{'
+        children = None
+        for k, v in val.items():
+            if k == 'children':
+                children = v
+                continue
+            dict_str += f'"{k}":{self.parse(v)},'
+        if children is not None:
+            dict_str += self.children_parser(children)
+        if dict_str.endswith(','):
+            dict_str = dict_str[:-1]
+        dict_str += '}'
+        return dict_str
+
+    def list_parser(self, val):
+        list_str = '['
+        for item in val:
+            item_str = self.parse(item)
+            list_str += item_str + ','
+        if list_str.endswith(','):
+            list_str = list_str[:-1]
+        list_str += ']'
+        return list_str
+
+    def int_parser(self, val):
+        return f'{val}'
+
+    def str_parser(self, val):
+        return json.dumps(val, ensure_ascii=False)
+
+    def bool_parser(self, val):
+        return json.dumps(val)
+
+    def none_parser(self, val):
+        return 'null'
+
+    def children_parser(self, val):
+        self.indent += 1
+        children_str = '"children":['
+        for child in val:
+            child_str = self.parse(child)
+            children_str += '\n' + '  '*self.indent + child_str + ','
+        if children_str.endswith(','):
+            children_str = children_str[:-1]
+        children_str += ']'
+        self.indent -= 1
+        return children_str
